ICMEM.

ICMEM was used to conduct spatial clustering with hidden Markov random field for fixed beta and fixed number of clusters

ICMEM(
  y,
  x_int,
  Adj,
  mu_int,
  sigma_int,
  alpha,
  beta_grid,
  PX,
  maxIter_ICM,
  maxIter
)

Arguments

y
is a matrix of PCs containing gene expression.

x_int
is a vector of initial cluster label.

Adj
is a matrix containing neighborhood information generated by find_neighbors2.

mu_int
is a initial mean vector. we often generated it by Gaussian mixture model.

sigma_int
is a initial co-variance matrix. we often generated it by Gaussian mixture model.

alpha
is a intercept.

beta_grid
is a smoothing parameter that can be specified by user.

PX
is a logical value specifying the parameter expansion in EM algorithm.

maxIter_ICM
is the maximum iteration of ICM algorithm.

maxIter
is the maximum iteration of EM algorithm.

Value

a list.

The item 'x' is the clustering result.

The item 'gam' is the posterior probability matrix.

The item 'ell' is the opposite log-likelihood.

The item 'mu' is the mean of each component.

The item 'sigma' is the variance of each component.

Details

ICMEM was used to conduct spatial clustering with hidden Markov random field for fixed beta and fixed number of clusters