selectKPlot.

The function selectKPlot is used to demonstrate the scatter plot of BIC or Modified BIC vs K for selecting the best K.

selectKPlot(SCobject, K_set = 2:10, criterion = "BIC", c = 1)

Arguments

SCobject
is a object generated from SC.MEB function.

K_set
is the corresponding K_set used in your previous function SC.MEB.

criterion
is a character specifying the criterion for selecting K. The default is BIC, the alternative criterion MBIC can also be used.

c
is a positive value in modified BIC. The default is 1. In general, For the ST or Visium data, it often ranges from 0.4 to 1 while for the MERFISH data with large number of cells, it often is larger. For example 10,20. Most importantly, SC-MEB is fast, scaling well in terms of sample size, which allow the user to tune the c based on their prior knowledge about the tissues or cells.

Value

a ggplot2 object.

Details

The function selectKPlot is used to demonstrate the scatter plot of BIC or Modified BIC vs K for selecting the best K.